mc_rate_control split out of mc_att_control
