+++ ROMFS/px4fmu_common/init.d/rc.vtol_apps
@@ -17,6 +17,7 @@ ekf2 start
 
 
 vtol_att_control start
+mc_rate_control start
 mc_att_control start
 mc_pos_control start
 fw_att_control start
