+++ boards/airmind/mindpx-v2/default.cmake
@@ -66,6 +66,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
