+++ boards/atlflight/excelsior/default.cmake
@@ -80,6 +80,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
