+++ boards/atlflight/excelsior/qurt-default.cmake
@@ -71,6 +71,7 @@ px4_add_board(
 		landing_target_estimator
 		local_position_estimator
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		sensors
 		sih
