+++ boards/auav/x21/default.cmake
@@ -72,6 +72,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
