+++ boards/bitcraze/crazyflie/default.cmake
@@ -33,6 +33,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
