+++ boards/emlid/navio2/cross.cmake
@@ -48,6 +48,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
