+++ boards/intel/aerofc-v1/default.cmake
@@ -49,6 +49,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
