+++ boards/intel/aerofc-v1/rtps.cmake
@@ -52,6 +52,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		micrortps_bridge
 		navigator
