+++ boards/nxp/fmuk66-v3/default.cmake
@@ -68,6 +68,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
