+++ boards/omnibus/f4sd/default.cmake
@@ -61,6 +61,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
