+++ boards/parrot/bebop/default.cmake
@@ -36,6 +36,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
