+++ boards/px4/fmu-v2/default.cmake
@@ -81,6 +81,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
