+++ boards/px4/fmu-v2/multicopter.cmake
@@ -54,6 +54,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
