+++ boards/px4/fmu-v3/stackcheck.cmake
@@ -79,6 +79,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
