+++ boards/px4/fmu-v4/rtps.cmake
@@ -66,6 +66,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		micrortps_bridge
 		navigator
