+++ boards/px4/fmu-v5/default.cmake
@@ -78,6 +78,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
