+++ boards/px4/fmu-v5/stackcheck.cmake
@@ -76,6 +76,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		#micrortps_bridge
 		navigator
