+++ boards/px4/fmu-v5x/multicopter.cmake
@@ -69,6 +69,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
