+++ boards/px4/raspberrypi/cross.cmake
@@ -41,6 +41,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
