+++ boards/px4/raspberrypi/native.cmake
@@ -39,6 +39,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		sensors
