+++ boards/px4/sitl/default.cmake
@@ -38,6 +38,7 @@ px4_add_board(
 		logger
 		mavlink
 		mc_att_control
+		mc_rate_control
 		mc_pos_control
 		navigator
 		replay
