+++ src/modules/mc_att_control/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2015-2019 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -36,8 +36,7 @@ add_subdirectory(AttitudeControl)
 px4_add_module(
 	MODULE modules__mc_att_control
 	MAIN mc_att_control
-	STACK_MAIN 1200
-	STACK_MAX 3500
+	STACK_MAIN 1500
 	COMPILE_FLAGS
 	SRCS
 		mc_att_control_main.cpp
