+++ src/modules/mc_att_control/mc_att_control_params.c
@@ -54,87 +54,6 @@
 PARAM_DEFINE_FLOAT(MC_ROLL_P, 6.5f);
 
 /**
- * Roll rate P gain
- *
- * Roll rate proportional gain, i.e. control output for angular speed error 1 rad/s.
- *
- * @min 0.0
- * @max 0.5
- * @decimal 3
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ROLLRATE_P, 0.15f);
-
-/**
- * Roll rate I gain
- *
- * Roll rate integral gain. Can be set to compensate static thrust difference or gravity center offset.
- *
- * @min 0.0
- * @decimal 3
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ROLLRATE_I, 0.2f);
-
-/**
- * Roll rate integrator limit
- *
- * Roll rate integrator limit. Can be set to increase the amount of integrator available to counteract disturbances or reduced to improve settling time after large roll moment trim changes.
- *
- * @min 0.0
- * @decimal 2
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_RR_INT_LIM, 0.30f);
-
-/**
- * Roll rate D gain
- *
- * Roll rate differential gain. Small values help reduce fast oscillations. If value is too big oscillations will appear again.
- *
- * @min 0.0
- * @max 0.01
- * @decimal 4
- * @increment 0.0005
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ROLLRATE_D, 0.003f);
-
-/**
- * Roll rate feedforward
- *
- * Improves tracking performance.
- *
- * @min 0.0
- * @decimal 4
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ROLLRATE_FF, 0.0f);
-
-/**
- * Roll rate controller gain
- *
- * Global gain of the controller.
- *
- * This gain scales the P, I and D terms of the controller:
- * output = MC_ROLLRATE_K * (MC_ROLLRATE_P * error
- * 			     + MC_ROLLRATE_I * error_integral
- * 			     + MC_ROLLRATE_D * error_derivative)
- * Set MC_ROLLRATE_P=1 to implement a PID in the ideal form.
- * Set MC_ROLLRATE_K=1 to implement a PID in the parallel form.
- *
- * @min 0.0
- * @max 5.0
- * @decimal 4
- * @increment 0.0005
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ROLLRATE_K, 1.0f);
-
-/**
  * Pitch P gain
  *
  * Pitch proportional gain, i.e. desired angular speed in rad/s for error 1 rad.
@@ -149,86 +68,6 @@ PARAM_DEFINE_FLOAT(MC_ROLLRATE_K, 1.0f);
 PARAM_DEFINE_FLOAT(MC_PITCH_P, 6.5f);
 
 /**
- * Pitch rate P gain
- *
- * Pitch rate proportional gain, i.e. control output for angular speed error 1 rad/s.
- *
- * @min 0.0
- * @max 0.6
- * @decimal 3
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_PITCHRATE_P, 0.15f);
-
-/**
- * Pitch rate I gain
- *
- * Pitch rate integral gain. Can be set to compensate static thrust difference or gravity center offset.
- *
- * @min 0.0
- * @decimal 3
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_PITCHRATE_I, 0.2f);
-
-/**
- * Pitch rate integrator limit
- *
- * Pitch rate integrator limit. Can be set to increase the amount of integrator available to counteract disturbances or reduced to improve settling time after large pitch moment trim changes.
- *
- * @min 0.0
- * @decimal 2
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_PR_INT_LIM, 0.30f);
-
-/**
- * Pitch rate D gain
- *
- * Pitch rate differential gain. Small values help reduce fast oscillations. If value is too big oscillations will appear again.
- *
- * @min 0.0
- * @decimal 4
- * @increment 0.0005
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_PITCHRATE_D, 0.003f);
-
-/**
- * Pitch rate feedforward
- *
- * Improves tracking performance.
- *
- * @min 0.0
- * @decimal 4
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_PITCHRATE_FF, 0.0f);
-
-/**
- * Pitch rate controller gain
- *
- * Global gain of the controller.
- *
- * This gain scales the P, I and D terms of the controller:
- * output = MC_PITCHRATE_K * (MC_PITCHRATE_P * error
- * 			     + MC_PITCHRATE_I * error_integral
- * 			     + MC_PITCHRATE_D * error_derivative)
- * Set MC_PITCHRATE_P=1 to implement a PID in the ideal form.
- * Set MC_PITCHRATE_K=1 to implement a PID in the parallel form.
- *
- * @min 0.0
- * @max 5.0
- * @decimal 4
- * @increment 0.0005
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_PITCHRATE_K, 1.0f);
-
-/**
  * Yaw P gain
  *
  * Yaw proportional gain, i.e. desired angular speed in rad/s for error 1 rad.
@@ -241,87 +80,6 @@ PARAM_DEFINE_FLOAT(MC_PITCHRATE_K, 1.0f);
  * @group Multicopter Attitude Control
  */
 PARAM_DEFINE_FLOAT(MC_YAW_P, 2.8f);
-
-/**
- * Yaw rate P gain
- *
- * Yaw rate proportional gain, i.e. control output for angular speed error 1 rad/s.
- *
- * @min 0.0
- * @max 0.6
- * @decimal 2
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_YAWRATE_P, 0.2f);
-
-/**
- * Yaw rate I gain
- *
- * Yaw rate integral gain. Can be set to compensate static thrust difference or gravity center offset.
- *
- * @min 0.0
- * @decimal 2
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_YAWRATE_I, 0.1f);
-
-/**
- * Yaw rate integrator limit
- *
- * Yaw rate integrator limit. Can be set to increase the amount of integrator available to counteract disturbances or reduced to improve settling time after large yaw moment trim changes.
- *
- * @min 0.0
- * @decimal 2
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_YR_INT_LIM, 0.30f);
-
-/**
- * Yaw rate D gain
- *
- * Yaw rate differential gain. Small values help reduce fast oscillations. If value is too big oscillations will appear again.
- *
- * @min 0.0
- * @decimal 2
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_YAWRATE_D, 0.0f);
-
-/**
- * Yaw rate feedforward
- *
- * Improves tracking performance.
- *
- * @min 0.0
- * @decimal 4
- * @increment 0.01
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_YAWRATE_FF, 0.0f);
-
-/**
- * Yaw rate controller gain
- *
- * Global gain of the controller.
- *
- * This gain scales the P, I and D terms of the controller:
- * output = MC_YAWRATE_K * (MC_YAWRATE_P * error
- * 			     + MC_YAWRATE_I * error_integral
- * 			     + MC_YAWRATE_D * error_derivative)
- * Set MC_YAWRATE_P=1 to implement a PID in the ideal form.
- * Set MC_YAWRATE_K=1 to implement a PID in the parallel form.
- *
- * @min 0.0
- * @max 5.0
- * @decimal 4
- * @increment 0.0005
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_YAWRATE_K, 1.0f);
 
 /**
  * Max roll rate
@@ -374,107 +132,6 @@ PARAM_DEFINE_FLOAT(MC_PITCHRATE_MAX, 220.0f);
 PARAM_DEFINE_FLOAT(MC_YAWRATE_MAX, 200.0f);
 
 /**
- * Max acro roll rate
- * default: 2 turns per second
- *
- * @unit deg/s
- * @min 0.0
- * @max 1800.0
- * @decimal 1
- * @increment 5
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ACRO_R_MAX, 720.0f);
-
-/**
- * Max acro pitch rate
- * default: 2 turns per second
- *
- * @unit deg/s
- * @min 0.0
- * @max 1800.0
- * @decimal 1
- * @increment 5
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ACRO_P_MAX, 720.0f);
-
-/**
- * Max acro yaw rate
- * default 1.5 turns per second
- *
- * @unit deg/s
- * @min 0.0
- * @max 1800.0
- * @decimal 1
- * @increment 5
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ACRO_Y_MAX, 540.0f);
-
-/**
- * Acro mode Expo factor for Roll and Pitch.
- *
- * Exponential factor for tuning the input curve shape.
- *
- * 0 Purely linear input curve
- * 1 Purely cubic input curve
- *
- * @min 0
- * @max 1
- * @decimal 2
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ACRO_EXPO, 0.69f);
-
-/**
- * Acro mode Expo factor for Yaw.
- *
- * Exponential factor for tuning the input curve shape.
- *
- * 0 Purely linear input curve
- * 1 Purely cubic input curve
- *
- * @min 0
- * @max 1
- * @decimal 2
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ACRO_EXPO_Y, 0.69f);
-
-/**
- * Acro mode SuperExpo factor for Roll and Pitch.
- *
- * SuperExpo factor for refining the input curve shape tuned using MC_ACRO_EXPO.
- *
- * 0 Pure Expo function
- * 0.7 resonable shape enhancement for intuitive stick feel
- * 0.95 very strong bent input curve only near maxima have effect
- *
- * @min 0
- * @max 0.95
- * @decimal 2
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ACRO_SUPEXPO, 0.7f);
-
-/**
- * Acro mode SuperExpo factor for Yaw.
- *
- * SuperExpo factor for refining the input curve shape tuned using MC_ACRO_EXPO_Y.
- *
- * 0 Pure Expo function
- * 0.7 resonable shape enhancement for intuitive stick feel
- * 0.95 very strong bent input curve only near maxima have effect
- *
- * @min 0
- * @max 0.95
- * @decimal 2
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_ACRO_SUPEXPOY, 0.7f);
-
-/**
  * Threshold for Rattitude mode
  *
  * Manual input needed in order to override attitude control rate setpoints
@@ -487,127 +144,6 @@ PARAM_DEFINE_FLOAT(MC_ACRO_SUPEXPOY, 0.7f);
  * @group Multicopter Attitude Control
  */
 PARAM_DEFINE_FLOAT(MC_RATT_TH, 0.8f);
-
-/**
- * Battery power level scaler
- *
- * This compensates for voltage drop of the battery over time by attempting to
- * normalize performance across the operating range of the battery. The copter
- * should constantly behave as if it was fully charged with reduced max acceleration
- * at lower battery percentages. i.e. if hover is at 0.5 throttle at 100% battery,
- * it will still be 0.5 at 60% battery.
- *
- * @boolean
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_INT32(MC_BAT_SCALE_EN, 0);
-
-/**
- * TPA P Breakpoint
- *
- * Throttle PID Attenuation (TPA)
- * Magnitude of throttle setpoint at which to begin attenuating roll/pitch P gain
- *
- * @min 0.0
- * @max 1.0
- * @decimal 2
- * @increment 0.1
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_TPA_BREAK_P, 1.0f);
-
-/**
- * TPA I Breakpoint
- *
- * Throttle PID Attenuation (TPA)
- * Magnitude of throttle setpoint at which to begin attenuating roll/pitch I gain
- *
- * @min 0.0
- * @max 1.0
- * @decimal 2
- * @increment 0.1
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_TPA_BREAK_I, 1.0f);
-
-/**
- * TPA D Breakpoint
- *
- * Throttle PID Attenuation (TPA)
- * Magnitude of throttle setpoint at which to begin attenuating roll/pitch D gain
- *
- * @min 0.0
- * @max 1.0
- * @decimal 2
- * @increment 0.1
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_TPA_BREAK_D, 1.0f);
-
-/**
- * TPA Rate P
- *
- * Throttle PID Attenuation (TPA)
- * Rate at which to attenuate roll/pitch P gain
- * Attenuation factor is 1.0 when throttle magnitude is below the setpoint
- * Above the setpoint, the attenuation factor is (1 - rate * (throttle - breakpoint) / (1.0 - breakpoint))
- *
- * @min 0.0
- * @max 1.0
- * @decimal 2
- * @increment 0.05
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_TPA_RATE_P, 0.0f);
-
-/**
- * TPA Rate I
- *
- * Throttle PID Attenuation (TPA)
- * Rate at which to attenuate roll/pitch I gain
- * Attenuation factor is 1.0 when throttle magnitude is below the setpoint
- * Above the setpoint, the attenuation factor is (1 - rate * (throttle - breakpoint) / (1.0 - breakpoint))
- *
- * @min 0.0
- * @max 1.0
- * @decimal 2
- * @increment 0.05
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_TPA_RATE_I, 0.0f);
-
-/**
- * TPA Rate D
- *
- * Throttle PID Attenuation (TPA)
- * Rate at which to attenuate roll/pitch D gain
- * Attenuation factor is 1.0 when throttle magnitude is below the setpoint
- * Above the setpoint, the attenuation factor is (1 - rate * (throttle - breakpoint) / (1.0 - breakpoint))
- *
- * @min 0.0
- * @max 1.0
- * @decimal 2
- * @increment 0.05
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_TPA_RATE_D, 0.0f);
-
-/**
- * Cutoff frequency for the low pass filter on the D-term in the rate controller
- *
- * The D-term uses the derivative of the rate and thus is the most susceptible to noise.
- * Therefore, using a D-term filter allows to decrease the driver-level filtering, which
- * leads to reduced control latency and permits to increase the P gains.
- * A value of 0 disables the filter.
- *
- * @unit Hz
- * @min 0
- * @max 1000
- * @decimal 0
- * @increment 10
- * @group Multicopter Attitude Control
- */
-PARAM_DEFINE_FLOAT(MC_DTERM_CUTOFF, 0.f);
 
 /**
  * Multicopter air-mode
@@ -626,4 +162,3 @@ PARAM_DEFINE_FLOAT(MC_DTERM_CUTOFF, 0.f);
  * @group Multicopter Attitude Control
  */
 PARAM_DEFINE_INT32(MC_AIRMODE, 0);
-
