+++ src/modules/mc_rate_control/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2018 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2019 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,5 +31,16 @@
 #
 ############################################################################
 
-px4_add_library(drivers_accelerometer PX4Accelerometer.cpp)
-target_link_libraries(drivers_accelerometer PRIVATE drivers__device mathlib)
+px4_add_module(
+	MODULE modules__mc_rate_control
+	MAIN mc_rate_control
+	STACK_MAIN 1500
+	COMPILE_FLAGS
+	SRCS
+		MulticopterRateControl.cpp
+	DEPENDS
+		circuit_breaker
+		conversion
+		mathlib
+		px4_work_queue
+	)
